requirements.txt
Flask==3.1.2
PyMuPDF==1.26.5
firebase-admin==6.5.0
python-dotenv==1.0.0
